TD 2 - Fixed email generation bugs caused by application settings refactoring. Fixed missing email section admin image.
--HG--
extra : convert_revision : svn%3A7c8e8105-a804-6c41-a338-bf07137d3968%40106

+++ TicketDesk-2-branch/TicketDesk.Domain/Services/Base/IApplicationSettingsService.cs
@@ -85,7 +85,7 @@ namespace TicketDesk.Domain.Services
         string EmailServiceName();
         double EmailNotificationInitialDelayMinutes();
         double EmailMaxConsolidationWaitMinutes();
-        double EmailResendDelayMinutes();
+        int EmailResendDelayMinutes();
         int EmailMaxDeliveryAttempts();
         string FromEmailDisplayName();
         string FromEmailAddress();
